+++ lib/routes.dart
@@ -460,6 +460,7 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
                   callRepository,
                   Get.find(),
                   userRepository,
+                  Get.find(),
                   me: me,
                 ),
               );
@@ -559,6 +560,7 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
                 callRepository,
                 Get.find(),
                 userRepository,
+                Get.find(),
                 me: me,
               ),
             );