+++ lib/ui/page/home/tab/chats/controller.dart
@@ -44,9 +44,11 @@ import '/domain/service/user.dart';
 import '/provider/gql/exceptions.dart'
     show
         CreateGroupChatException,
+        FavoriteChatException,
         HideChatException,
         RemoveChatMemberException,
-        ToggleChatMuteException;
+        ToggleChatMuteException,
+        UnfavoriteChatException;
 import '/routes.dart';
 import '/ui/widget/text_field.dart';
 import '/util/message_popup.dart';
@@ -489,6 +491,30 @@ class ChatsTabController extends GetxController {
     }
   }
 
+  /// Marks the specified [Chat] identified by its [id] as favorited.
+  Future<void> favoriteChat(ChatId id) async {
+    try {
+      await _chatService.favoriteChat(id);
+    } on FavoriteChatException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    }
+  }
+
+  /// Removes the specified [Chat] identified by its [id] from the favorites.
+  Future<void> unfavoriteChat(ChatId id) async {
+    try {
+      await _chatService.unfavoriteChat(id);
+    } on UnfavoriteChatException catch (e) {
+      MessagePopup.error(e);
+    } catch (e) {
+      MessagePopup.error(e);
+      rethrow;
+    }
+  }
+
   /// Returns an [User] from [UserService] by the provided [id].
   Future<RxUser?> getUser(UserId id) => _userService.get(id);
 
@@ -501,6 +527,18 @@ class ChatsTabController extends GetxController {
       } else if (a.chat.value.ongoingCall == null &&
           b.chat.value.ongoingCall != null) {
         return 1;
+      }
+
+      if (a.chat.value.favoritePosition != null &&
+          b.chat.value.favoritePosition == null) {
+        return -1;
+      } else if (a.chat.value.favoritePosition == null &&
+          b.chat.value.favoritePosition != null) {
+        return 1;
+      } else if (a.chat.value.favoritePosition != null &&
+          b.chat.value.favoritePosition != null) {
+        return a.chat.value.favoritePosition!
+            .compareTo(b.chat.value.favoritePosition!);
       }
 
       return b.chat.value.updatedAt.compareTo(a.chat.value.updatedAt);