+++ lib/ui/page/home/tab/chats/widget/recent_chat.dart
@@ -18,7 +18,6 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
-import '/ui/page/home/widget/retry_image.dart';
 import '/api/backend/schema.dart' show ChatMemberInfoAction;
 import '/domain/model/attachment.dart';
 import '/domain/model/chat.dart';
@@ -37,7 +36,7 @@ import '/ui/page/home/tab/chats/widget/periodic_builder.dart';
 import '/ui/page/home/widget/animated_typing.dart';
 import '/ui/page/home/widget/avatar.dart';
 import '/ui/page/home/widget/chat_tile.dart';
-import '/ui/page/home/widget/init_callback.dart';
+import '/ui/page/home/widget/retry_image.dart';
 import '/ui/widget/context_menu/menu.dart';
 import '/ui/widget/svg/svg.dart';
 import '/ui/widget/widget_button.dart';
@@ -57,6 +56,8 @@ class RecentChatTile extends StatelessWidget {
     this.onJoin,
     this.onMute,
     this.onUnmute,
+    this.onFavorite,
+    this.onUnfavorite,
   }) : super(key: key);
 
   /// [RxChat] this [RecentChatTile] is about.
@@ -92,6 +93,13 @@ class RecentChatTile extends StatelessWidget {
 
   /// Callback, called when this [rxChat] unmute action is triggered.
   final void Function()? onUnmute;
+
+  /// Callback, called when this [rxChat] add to favorites action is triggered.
+  final void Function()? onFavorite;
+
+  /// Callback, called when this [rxChat] remove from favorites action is
+  /// triggered.
+  final void Function()? onUnfavorite;
 
   @override
   Widget build(BuildContext context) {
@@ -138,6 +146,17 @@ class RecentChatTile extends StatelessWidget {
           ),
         ],
         actions: [
+          chat.favoritePosition != null
+              ? ContextMenuButton(
+                  key: const Key('UnfavoriteChatButton'),
+                  label: 'btn_delete_from_favorites'.l10n,
+                  onPressed: onUnfavorite,
+                )
+              : ContextMenuButton(
+                  key: const Key('FavoriteChatButton'),
+                  label: 'btn_add_to_favorites'.l10n,
+                  onPressed: onFavorite,
+                ),
           ContextMenuButton(
             key: const Key('ButtonHideChat'),
             label: 'btn_hide_chat'.l10n,