+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All user visible changes to this project will be documented in this file. This p
         - Drafts. ([#173], [#102])
     - Chats tab:
         - Chat muting/unmuting. ([#172], [#63])
+        - Favorite chats. ([#218], [#209])
     - Home page:
         - Quick status changing menu. ([#204], [#203])
 
@@ -29,6 +30,10 @@ All user visible changes to this project will be documented in this file. This p
         - Redesigned chats tab. ([#142])
     - Media panel:
         - Video resizing when dragged. ([#191], [#190])
+    - Chat page:
+        - Redesigned gallery. ([#212], [#199])
+    - Chats tab:
+        - Redesigned attachments preview. ([#217], [#214])
 
 ### Fixed
 
@@ -54,8 +59,14 @@ All user visible changes to this project will be documented in this file. This p
 [#193]: /../../pull/193
 [#196]: /../../issues/196
 [#198]: /../../pull/198
+[#199]: /../../issues/199
 [#203]: /../../issues/203
 [#204]: /../../pull/204
+[#209]: /../../issues/209
+[#212]: /../../pull/212
+[#214]: /../../issues/214
+[#217]: /../../pull/217
+[#218]: /../../pull/218